DDP-7047 | introduce BaseGenerator which will separate Mapping and Source generators
