+++ src/main/java/org/broadinstitute/dsm/model/elastic/export/MappingGenerator.java
@@ -1,14 +1,40 @@
 package org.broadinstitute.dsm.model.elastic.export;
 
+import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
+import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.dsm.model.NameValue;
 
-public class MappingGenerator implements Generator {
+public class MappingGenerator extends BaseGenerator {
 
 
     @Override
     public Map<String, Object> generate(NameValue nameValue) {
+        initializeNecessaryFields(Objects.requireNonNull(nameValue));
+        String property = TABLE_ALIAS_MAPPINGS.get(dbElement.getTableAlias());
+        String value = (String) nameValue.getValue();
+
+        StringUtils.isNumeric("asd");
+        Boolean.valueOf("true");
+        // Date, time, datetime
+        
+
+
+        Map<String, Object> jsonMap = new HashMap<>();
+        Map<String, Object> message = new HashMap<>();
+        message.put("type", "");
+        Map<String, Object> properties = new HashMap<>();
+        properties.put("message", message);
+        jsonMap.put("properties", properties);
+        request.source(jsonMap);
+
+
+
         return null;
     }
+
+
+
 }
